So...step works. Had to do a few minor adjustments to ensure smooth operation. First, we created an IRQ to handle execution of a program (IE: run <pid>)... that will ensure no matter what our program not only makes it to the readyqueue, but gets passed to the function to handle a PCB (krnProcess)... inside krnProcess we set isExecuting to the opposite of a global variable _StepCPU which gets "turned on" via the shell command step <on | off>. So, with isExecuting set to false, cpu.cycle() does not run, which allows us to call cpu.cycle() with each click of the "step" button by firing another IRQ which calls cpu.cycle().
so --> a lot has happened here and it made me think of how to handle the
"run" command...Glad I saved step for last, I really am getting to know
my way around now:)

Next commit will have 1 minor issue fixed. We are not telling the PCB
that it is terminated. Once terminated, we do not want the ability to
call it again! BAD ... but leave it in our list for later. (maybe review
or whatever _ ram is cheap).