step is still a bit buggy. I added the ability to keep the current state of the PCB up to date by setting it in various locations. To New after load, to ready after run (although that is so breif since it runs asap and flips to running) to waiting if an interrupt comes in and to running during a cpu cycle.
more to come on that buggy step. like it works, but I do not want it to
continue to step through a program once we are done with it. the program
should have to be loaded all over again... I feel my current working
version is worth 7 or 8 points though :).