added a function in utils that checks the size of my readyqueue by returning a number of how many non-terminated processes are contained within.
also fixed it to not allow more processes than the system can handle at
one time (currently 23 - one example from class website did 24 but 25
would crash the OS... must be a difference in swapping, for 1 more
process it is not worth the headache.

I also removed some comments that no longer pertained -- in schedule

1 more minor bug to fix, when a process is taken from the residentqueue
to the readyqueue the pcb state is not updated _ it does not get updated
until it actually runs...that causes the ability to type run or runall
again and it actually working!